Create Rust application connecting to Reddit API and self updating
The program fetches data from a list of subreddits every 15 minutes and updates the information such as karma, comments, posts. This allows for a tracking how each comment or post is getting traction and so on. The data is being stored on a postgresql server which should allow for easy access by any other application.

 Date:      Tue Nov 16 13:57:18 2021 +0100
	new file:   Cargo.lock
	new file:   Cargo.toml
	new file:   src/main.rs